+++ CHANGELOG.md
@@ -10,6 +10,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `smartcast` can now disable splitting by specifying `allow_split=False`
 * ### Broken: Added `required` as an option to `Value`.
 
+* The `short_alias` keyword to `Value`, which allows single '-' prefix chars on
+  the command line.
+
+* The `port_argparse` method which takes an existing (simple) argparse CLI and
+  attempts to make a scriptconfig version of it.
+
+### Changed
+* The `alias` keyword no expands a single argparse argument instead of making
+  multiple of them.
+
 
 ## Version 0.5.8 - Released 2021-05-19